Complete ABMS

+++ src/main/java/juamlosada/challengebesysoft/services/impl/ProductServiceImpl.java
@@ -1,19 +1,26 @@
 package juamlosada.challengebesysoft.services.impl;
 
+import juamlosada.challengebesysoft.configuration.exception.duplicateentry.DuplicateProductException;
+import juamlosada.challengebesysoft.configuration.exception.notfound.CategoryNotFoundException;
 import juamlosada.challengebesysoft.configuration.exception.notfound.ProductNotFoundException;
+import juamlosada.challengebesysoft.entities.Category;
 import juamlosada.challengebesysoft.entities.Product;
+import juamlosada.challengebesysoft.repositories.CategoryRepository;
 import juamlosada.challengebesysoft.repositories.ProductRepository;
 import juamlosada.challengebesysoft.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ProductServiceImpl implements ProductService {
 
     @Autowired
     private ProductRepository productRepository;
+    @Autowired
+    private CategoryRepository categoryRepository;
 
     @Override
     public Product getById(Long id) {
@@ -22,7 +29,18 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public Product save(Product product) {
+        if(productRepository.existsByName(product.getName())) {
+            throw new DuplicateProductException(product.getName());
+        }
+
+        Category category = categoryRepository.findById(product.getCategory().getId()).orElseThrow(() -> new CategoryNotFoundException(product.getCategory().getId()));
+
+        product.setCategory(category);
+
+        category.getProducts().add(product);
+
         return productRepository.save(product);
+
     }
 
     @Override
@@ -32,6 +50,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteById(Long id) {
+        if(!productRepository.existsById(id)) {
+            throw new ProductNotFoundException(id);
+        }
         productRepository.deleteById(id);
     }
 
@@ -40,9 +61,22 @@ public class ProductServiceImpl implements ProductService {
 
         Product prod = productRepository.findById(id).orElseThrow(() -> new ProductNotFoundException(id));
 
+        if(productRepository.existsByName(product.getName()) && !Objects.equals(prod.getName(), product.getName())) {
+            throw new DuplicateProductException(product.getName());
+        }
+
+
+        Category categoryNew = categoryRepository.findById(product.getCategory().getId()).orElseThrow(() -> new CategoryNotFoundException(product.getCategory().getId()));
+
+
+        if (!categoryNew.getProducts().contains(prod)) {
+            categoryNew.getProducts().add(prod);
+            prod.getCategory().getProducts().remove(prod);
+        }
+
         prod.setName(product.getName());
         prod.setPrice(product.getPrice());
-        prod.setCategory(product.getCategory());
+        prod.setCategory(categoryNew);
         prod.setSales(product.getSales());
 
         return productRepository.save(prod);
